+++ csscheme.py
@@ -1,6 +1,4 @@
-import re
 import os
-import subprocess
 
 import sublime
 import sublime_plugin
@@ -8,121 +6,97 @@ import sublime_plugin
 try:
     # Use a different name because PackageDev adds it to the path and that
     # takes precedence over local paths (for some reason).
-    from .my_sublime_lib.view import OutputPanel
     from .my_sublime_lib import WindowAndTextCommand
-except:
-    from my_sublime_lib.view import OutputPanel
-    from my_sublime_lib import WindowAndTextCommand
+    from .my_sublime_lib.view import OutputPanel
 
-try:
-    from .tinycsscheme.parser import CSSchemeParser, ParseError
+    from .tinycsscheme.parser import CSSchemeParser
     from .tinycsscheme.dumper import CSSchemeDumper, DumpError
+
     from .scope_data import COMPILED_HEADS
+    from . import converters
 except:
-    from tinycsscheme.parser import CSSchemeParser, ParseError
+    from my_sublime_lib import WindowAndTextCommand
+    from my_sublime_lib.view import OutputPanel
+
+    from tinycsscheme.parser import CSSchemeParser
     from tinycsscheme.dumper import CSSchemeDumper, DumpError
+
     from scope_data import COMPILED_HEADS
+    import converters
 
 
-# Returns a function for use with `re.sub`, requires matches in groups 1 and 2
-def swap_path_line(pattern, rel_dir):
-    def repl(m):
-        # Make path relative because we don't need long paths if in same dir
-        path = os.path.relpath(m.group(2), rel_dir)
-        return pattern % (path, m.group(1))
-    return repl
+###############################################################################
+
+
+PACKAGE = "CSScheme"  # my_sublime_lib.path.get_package_name() # __package__
 
 
 def settings():
+    """Load the settings file."""
     # We can safely call this over and over because it caches internally
     return sublime.load_settings("CSScheme.sublime-settings")
 
 
-def status(msg):
-    package = "CSScheme"  # my_sublime_lib.path.get_package_name()
-    sublime.status_message("%s: %s" % (package, msg))
-    print("[%s] %s" % (package, msg))
+def status(msg, printonly=""):
+    """Show a message in the statusbar and print to the console."""
+    sublime.status_message("%s: %s" % (PACKAGE, msg))
+    if printonly:
+        msg = msg + '\n' + printonly
+    print("[%s] %s" % (PACKAGE, msg))
 
 
-# Use window (and text) command to be able to call this command from both sources
-# (build systems are always window commands).
+###############################################################################
+
+
+# Use window (and text) command to be able to call this command from both
+# sources (build systems are always window commands).
 class convert_csscheme(WindowAndTextCommand):
-    """docs
-    """
-    def is_enabled(self):
-        return self.get_in_ext() is not None
 
-    def get_in_ext(self):
-        if not self.view.file_name():
-            return None
-        m = re.search(r'\.((?:sc|sa|c)ss)cheme$', self.view.file_name())
-        return m and m.group(1)
+    """Convert the active CSScheme (or variant) file into a .tmTheme plist."""
+
+    def is_enabled(self):
+        path = self.view.file_name()
+        return bool(path) and any(conv.valid_file(path) for conv in converters.all)
 
     def run(self, edit=None):
         if self.view.is_dirty():
-            return status("Save the file first")
+            return status("Save the file first.")
 
         in_file = self.view.file_name()
-        in_ext = self.get_in_ext()
         in_dir, in_base = os.path.split(in_file)
         out_file = os.path.splitext(in_file)[0] + '.tmTheme'
 
-        # TODO do this in oop style (with error reporting and cmd and stuff)
-        # Will probably do when more convertion options are added, if at all (syntaxes are a pita)
-        sass_path = settings().get('sass_path', 'sass')
-        commands = dict(
-            sass=[sass_path, '-l'],
-            scss=[sass_path, '-l',  '--scss'],
-            # less='less',
-            # stylus= ...
-        )
-
         # Open up output panel and auto-finalize it when we are done
         with OutputPanel(self.view.window(), "csscheme") as out:
+
+            # Determine our converter
+            conv = tuple(c for c in converters.all if c.valid_file(in_file))
+            if len(conv) > 1:
+                print(conv)
+                out.write_line("Found multiple contenders for conversion.\n"
+                               "If this happened to you, please tell the developer "
+                               "(me) to add code for this case. Thanks.")
+                return
+            assert len(conv) == 1
+            conv = conv[0]
+
             out.set_path(in_dir)
-            text = ""
-            if in_ext in commands:
-                try:
-                    process = subprocess.Popen(commands[in_ext] + [in_file],
-                                               stdout=subprocess.PIPE,
-                                               stderr=subprocess.PIPE,
-                                               shell=sublime.platform() == 'windows',
-                                               universal_newlines=True)
-                    text, stderr = process.communicate()
-                except Exception as e:
-                    out.write_line("Error converting from %s to CSS:\n"
-                                   "%s: %s" % (in_ext, e.__class__.__name__, e))
-                    return
+            executables = settings().get("executables")  # TOTEST
 
-                if process.returncode:
-                    out.set_regex(r"^\s+in (.*?) on line (\d+)$")
+            # Run converter
+            text = conv.convert(out, in_file, executables)
+            if not text:
+                return
 
-                    out.write_line("Errors converting from %s to CSS, return code: %s\n"
-                                   % (in_ext, process.returncode))
-                    # Swap line and path because sublime can't parse them otherwise
-                    out.write_line(re.sub(r"on line (\d+) of (.*?)$",
-                                          swap_path_line(r"in %s on line %s", in_dir),
-                                          stderr,
-                                          flags=re.M))
-                    return
-
-                elif not text:
-                    out.write_line("Unexpected error converting from %s to CSS:\nNo output"
-                                   % in_ext)
-                    return
-
-                elif stderr:
-                    out.write_line(stderr + "\n")
-            else:
-                assert in_ext == 'css'
-                text = self.view.substr(sublime.Region(0, self.view.size()))
-
-            # DEBUG
-            if settings().get('preview_compiled_css') and in_ext != 'css':
+            # Preview converted css for debugging, optionally
+            if settings().get('preview_compiled_css') and not conv.ext == 'csscheme':
                 v = self.view.window().new_file()
                 v.set_scratch(True)
-                v.set_syntax_file("Packages/CSScheme/CSScheme.tmLanguage")
-                from my_sublime_lib.edit import Edit
+                v.set_syntax_file("Packages/%s/CSScheme.tmLanguage" % PACKAGE)
+                try:
+                    from .my_sublime_lib.edit import Edit
+                except:
+                    from my_sublime_lib.edit import Edit
                 with Edit(v) as edit:
                     edit.append(text)
 
@@ -131,49 +105,7 @@ class convert_csscheme(WindowAndTextCommand):
 
             # Do some awesome error printing action
             if stylesheet.errors:
-                if in_ext in ('sass', 'scss'):
-                    err_reg = re.compile(r"/\* line (\d+), (.*?) \*/", re.M)
-
-                    # Match our modified output
-                    out.set_regex(r"^\s*/\* (.*?), line (\d+) \*/")
-
-                    lines = text.split('\n')
-                    # I could wrap this in an Edit(out.view) call because I modify it so often
-                    for e in stylesheet.errors:
-                        assert isinstance(e, ParseError)
-                        out.write_line("ParseError from CSS on line %d:" % e.line)
-
-                        # Search for last known line number (max 20)
-                        start_dump = 0
-                        for i in range(e.line, e.line - 20, -1):
-                            if i < 0:
-                                break
-                            m = re.match(r"\s*/\* line (\d+),", lines[i])
-                            if not m:
-                                continue
-                            start_dump = i
-                            # Swap line and path because sublime can't parse them otherwise
-                            out.write_line(
-                                "  " + err_reg.sub(swap_path_line("/* %s, line %s */", in_dir),
-                                                   lines[i])
-                            )
-                            break
-
-                        # Nothing found in the past lines, just print the erroneous line then
-                        if not start_dump:
-                            start_dump = e.line - 2
-
-                        for i in range(start_dump + 1, e.line):
-                            out.write_line("  " + lines[i])
-                        # Mark the column where the error happened (since we don't have source code)
-                        out.write_line("  %s^" % ('-' * (e.column - 1)))
-                        out.write_line("%s\n" % (e.reason))
-
-                elif in_ext == 'css':
-                    out.set_regex(r"^(.*):(\d+):(\d+):$")
-                    for e in stylesheet.errors:
-                        out.write_line("%s:%s:%s:\n  %s\n"
-                                       % (in_base, e.line, e.column, e.reason))
+                conv.report_parse_errors(out, in_file, text, stylesheet.errors)
                 return
             elif not stylesheet.rules:
                 # The CSS seems to be ... empty?
@@ -184,15 +116,7 @@ class convert_csscheme(WindowAndTextCommand):
             try:
                 CSSchemeDumper().dump_stylesheet_file(out_file, stylesheet)
             except DumpError as e:
-                if in_ext == 'css':
-                    out.set_regex(r"^(.*):(\d+):(\d+):$")
-                    out.write_line("%s:%s:%s:\n  %s%s\n"
-                                   % (in_base, e.line, e.column, e.reason, e.location))
-                else:
-                    # We can't accurately determine where the error occured (besides searching for
-                    # the last referenced number like above, and that kinda sucks), so just use text
-                    out.write_line("Error in data:\n  %s%s\n" % (e.reason, e.location))
-                return
+                conv.report_dump_error(out, in_file, text, e)
 
             status("Build successful")
             # Open out_file
@@ -203,7 +127,6 @@ class convert_csscheme(WindowAndTextCommand):
 ###############################################################################
 
 
-# TODO completions (scope_data from PackageDev in selectors, known properties (from dumper))
 class CSSchemeCompletionListener(sublime_plugin.EventListener):
     def __init__(self):
         properties = []
@@ -225,7 +148,6 @@ class CSSchemeCompletionListener(sublime_plugin.EventListener):
 
         match_sel = lambda s: all(view.match_selector(l, s) for l in locations)
 
-        # import spdb ; spdb.start()
         # Check context
         if not match_sel("source.csscheme - comment - string - variable"):
             return
@@ -239,10 +161,9 @@ class CSSchemeCompletionListener(sublime_plugin.EventListener):
 
         scope = self.get_scope(view, locations[0])
 
-        # We can't work with different scopes
-        for l in locations:
-            if self.get_scope(view, l) != scope:
-                return
+        # We can't work with different prefixes
+        if any(self.get_scope(view, l) != scope for l in locations):
+            return
 
         # Tokenize the current selector (only to the cursor)
         tokens = scope.split(".")